refactor: tidy the stuffs of input method service
